WetHumidGasPvt: move implementation to impl file
